move config related functions to their own class
