--- src/baseConfig.ts
@@ -1,9 +1,0 @@
-import { FilePath } from './cli';
-
-export const generateTempConfig = (existingConfigPath: FilePath, files: FilePath[]): string => {
-    return JSON.stringify({
-        extends: existingConfigPath,
-        files,
-        include: []
-    });
-};
